+++ lib/gtk.dart
@@ -1,4 +1,5 @@
 library gtk;
 
+export 'src/models/models.dart';
 export 'src/utils/utils.dart';
 export 'src/widgets/widgets.dart';