+++ lib/src/models/models.dart
@@ -1,0 +1,2 @@
+export 'view_switcher_data.dart';
+export 'view_switcher_style.dart';