+++ lib/src/models/view_switcher_style.dart
@@ -1,0 +1,1 @@
+enum ViewSwitcherStyle { desktop, mobile }